FossilOrigin-Name: 44bed8535436c7118007a5dc87d9cfa95edb1489

+++ src/add.c
@@ -25,6 +25,24 @@
 #include "cygsup.h"
 
 /*
+** WARNING: For Fossil version 1.x this value was always zero.  For Fossil
+**          2.x, it will probably always be one.  When this value is zero,
+**          files in the checkout will not be moved by the "mv" command and
+**          files in the checkout will not be removed by the "rm" command.
+**
+**          If the FOSSIL_ENABLE_LEGACY_MV_RM compile-time option is used,
+**          the "mv-rm-files" setting will be consulted instead of using
+**          this value.
+**
+**          To retain the Fossil version 1.x behavior when using Fossil 2.x,
+**          the FOSSIL_ENABLE_LEGACY_MV_RM compile-time option must be used
+**          -AND- the "mv-rm-files" setting must be set to zero.
+*/
+#ifndef FOSSIL_MV_RM_FILE
+#define FOSSIL_MV_RM_FILE                        (0)
+#endif
+
+/*
 ** This routine returns the names of files in a working checkout that
 ** are created by Fossil itself, and hence should not be added, deleted,
 ** or merge, and should be omitted from "clean" and "extras" lists.
@@ -274,6 +292,7 @@ void add_cmd(void){
   if( zIgnoreFlag==0 ){
     zIgnoreFlag = db_get("ignore-glob", 0);
   }
+  if( db_get_boolean("dotfiles", 0) ) scanFlags |= SCAN_ALL;
   vid = db_lget_int("checkout",0);
   db_begin_transaction();
   db_multi_exec("CREATE TEMP TABLE sfile(x TEXT PRIMARY KEY %s)",
@@ -334,32 +353,115 @@ void add_cmd(void){
 }
 
 /*
-** COMMAND: rm
-** COMMAND: delete*
+** This function adds a file to list of files to delete from disk after
+** the other actions required for the parent operation have completed
+** successfully.  The first time it is called for the current process,
+** it creates a temporary table named "fremove", to keep track of these
+** files.
+*/
+static void add_file_to_remove(
+  const char *zOldName /* The old name of the file on disk. */
+){
+  static int tableCreated = 0;
+  Blob fullOldName;
+  if( !tableCreated ){
+    db_multi_exec("CREATE TEMP TABLE fremove(x TEXT PRIMARY KEY %s)",
+                  filename_collation());
+    tableCreated = 1;
+  }
+  file_canonical_name(zOldName, &fullOldName, 0);
+  db_multi_exec("INSERT INTO fremove VALUES('%q');", blob_str(&fullOldName));
+  blob_reset(&fullOldName);
+}
+
+/*
+** This function deletes files from the checkout, using the file names
+** contained in the temporary table "fremove".  The temporary table is
+** created on demand by the add_file_to_remove() function.
 **
-** Usage: %fossil rm FILE1 ?FILE2 ...?
-**    or: %fossil delete FILE1 ?FILE2 ...?
+** If dryRunFlag is non-zero, no files will be removed; however, their
+** names will still be output.
+**
+** The temporary table "fremove" is dropped after being processed.
+*/
+static void process_files_to_remove(
+  int dryRunFlag /* Zero to actually operate on the file-system. */
+){
+  Stmt remove;
+  db_prepare(&remove, "SELECT x FROM fremove ORDER BY x;");
+  while( db_step(&remove)==SQLITE_ROW ){
+    const char *zOldName = db_column_text(&remove, 0);
+    if( !dryRunFlag ){
+      file_delete(zOldName);
+    }
+    fossil_print("DELETED_FILE %s\n", zOldName);
+  }
+  db_finalize(&remove);
+  db_multi_exec("DROP TABLE fremove;");
+}
+
+/*
+** COMMAND: rm
+** COMMAND: delete
+** COMMAND: forget*
+**
+** Usage: %fossil rm|delete|forget FILE1 ?FILE2 ...?
 **
 ** Remove one or more files or directories from the repository.
 **
-** This command does NOT remove the files from disk.  It just marks the
-** files as no longer being part of the project.  In other words, future
-** changes to the named files will not be versioned.
+** The 'rm' and 'delete' commands do NOT normally remove the files from
+** disk.  They just mark the files as no longer being part of the project.
+** In other words, future changes to the named files will not be versioned.
+** However, the default behavior of this command may be overridden via the
+** command line options listed below and/or the 'mv-rm-files' setting.
+**
+** The 'forget' command never removes files from disk, even when the command
+** line options and/or the 'mv-rm-files' setting would otherwise require it
+** to do so.
+**
+** WARNING: If the "--hard" option is specified -OR- the "mv-rm-files"
+**          setting is non-zero, files WILL BE removed from disk as well.
+**          This does NOT apply to the 'forget' command.
 **
 ** Options:
+**   --soft                  Skip removing files from the checkout.
+**                           This supersedes the --hard option.
+**   --hard                  Remove files from the checkout.
 **   --case-sensitive <BOOL> Override the case-sensitive setting.
+**   -n|--dry-run            If given, display instead of run actions.
 **
 ** See also: addremove, add
 */
 void delete_cmd(void){
   int i;
+  int removeFiles;
+  int dryRunFlag;
+  int softFlag;
+  int hardFlag;
   Stmt loop;
+
+  dryRunFlag = find_option("dry-run","n",0)!=0;
+  softFlag = find_option("soft",0,0)!=0;
+  hardFlag = find_option("hard",0,0)!=0;
 
   /* We should be done with options.. */
   verify_all_options();
 
   db_must_be_within_tree();
   db_begin_transaction();
+  if( g.argv[1][0]=='f' ){ /* i.e. "forget" */
+    removeFiles = 0;
+  }else if( softFlag ){
+    removeFiles = 0;
+  }else if( hardFlag ){
+    removeFiles = 1;
+  }else{
+#if FOSSIL_ENABLE_LEGACY_MV_RM
+    removeFiles = db_get_boolean("mv-rm-files",0);
+#else
+    removeFiles = FOSSIL_MV_RM_FILE;
+#endif
+  }
   db_multi_exec("CREATE TEMP TABLE sfile(x TEXT PRIMARY KEY %s)",
                 filename_collation());
   for(i=2; i<g.argc; i++){
@@ -383,13 +485,17 @@ void delete_cmd(void){
   db_prepare(&loop, "SELECT x FROM sfile");
   while( db_step(&loop)==SQLITE_ROW ){
     fossil_print("DELETED %s\n", db_column_text(&loop, 0));
+    if( removeFiles ) add_file_to_remove(db_column_text(&loop, 0));
   }
   db_finalize(&loop);
-  db_multi_exec(
-    "UPDATE vfile SET deleted=1 WHERE pathname IN sfile;"
-    "DELETE FROM vfile WHERE rid=0 AND deleted;"
-  );
+  if( !dryRunFlag ){
+    db_multi_exec(
+      "UPDATE vfile SET deleted=1 WHERE pathname IN sfile;"
+      "DELETE FROM vfile WHERE rid=0 AND deleted;"
+    );
+  }
   db_end_transaction(0);
+  if( removeFiles ) process_files_to_remove(dryRunFlag);
 }
 
 /*
@@ -535,6 +641,7 @@ void addremove_cmd(void){
   if( zIgnoreFlag==0 ){
     zIgnoreFlag = db_get("ignore-glob", 0);
   }
+  if( db_get_boolean("dotfiles", 0) ) scanFlags |= SCAN_ALL;
   vid = db_lget_int("checkout",0);
   db_begin_transaction();
 
@@ -593,7 +700,8 @@ void addremove_cmd(void){
 static void mv_one_file(
   int vid,
   const char *zOrig,
-  const char *zNew
+  const char *zNew,
+  int dryRunFlag
 ){
   int x = db_int(-1, "SELECT deleted FROM vfile WHERE pathname=%Q %s",
                          zNew, filename_collation());
@@ -607,10 +715,71 @@ static void mv_one_file(
     }
   }
   fossil_print("RENAME %s %s\n", zOrig, zNew);
-  db_multi_exec(
-    "UPDATE vfile SET pathname='%q' WHERE pathname='%q' %s AND vid=%d",
-    zNew, zOrig, filename_collation(), vid
-  );
+  if( !dryRunFlag ){
+    db_multi_exec(
+      "UPDATE vfile SET pathname='%q' WHERE pathname='%q' %s AND vid=%d",
+      zNew, zOrig, filename_collation(), vid
+    );
+  }
+}
+
+/*
+** This function adds a file to list of files to move on disk after the
+** other actions required for the parent operation have completed
+** successfully.  The first time it is called for the current process,
+** it creates a temporary table named "fmove", to keep track of these
+** files.
+*/
+static void add_file_to_move(
+  const char *zOldName, /* The old name of the file on disk. */
+  const char *zNewName  /* The new name of the file on disk. */
+){
+  static int tableCreated = 0;
+  Blob fullOldName;
+  Blob fullNewName;
+  if( !tableCreated ){
+    db_multi_exec("CREATE TEMP TABLE fmove(x TEXT PRIMARY KEY %s, y TEXT %s)",
+                  filename_collation(), filename_collation());
+    tableCreated = 1;
+  }
+  file_canonical_name(zOldName, &fullOldName, 0);
+  file_canonical_name(zNewName, &fullNewName, 0);
+  db_multi_exec("INSERT INTO fmove VALUES('%q','%q');",
+                blob_str(&fullOldName), blob_str(&fullNewName));
+  blob_reset(&fullNewName);
+  blob_reset(&fullOldName);
+}
+
+/*
+** This function moves files within the checkout, using the file names
+** contained in the temporary table "fmove".  The temporary table is
+** created on demand by the add_file_to_move() function.
+**
+** If dryRunFlag is non-zero, no files will be moved; however, their
+** names will still be output.
+**
+** The temporary table "fmove" is dropped after being processed.
+*/
+static void process_files_to_move(
+  int dryRunFlag /* Zero to actually operate on the file-system. */
+){
+  Stmt move;
+  db_prepare(&move, "SELECT x, y FROM fmove ORDER BY x;");
+  while( db_step(&move)==SQLITE_ROW ){
+    const char *zOldName = db_column_text(&move, 0);
+    const char *zNewName = db_column_text(&move, 1);
+    if( !dryRunFlag ){
+      if( file_wd_islink(zOldName) ){
+        symlink_copy(zOldName, zNewName);
+      }else{
+        file_copy(zOldName, zNewName);
+      }
+      file_delete(zOldName);
+    }
+    fossil_print("MOVED_FILE %s\n", zOldName);
+  }
+  db_finalize(&move);
+  db_multi_exec("DROP TABLE fmove;");
 }
 
 /*
@@ -623,23 +792,44 @@ static void mv_one_file(
 ** Move or rename one or more files or directories within the repository tree.
 ** You can either rename a file or directory or move it to another subdirectory.
 **
-** This command does NOT rename or move the files on disk.  This command merely
-** records the fact that filenames have changed so that appropriate notations
-** can be made at the next commit/checkin.
+** The 'mv' command does NOT normally rename or move the files on disk.
+** This command merely records the fact that file names have changed so
+** that appropriate notations can be made at the next commit/check-in.
+** However, the default behavior of this command may be overridden via
+** command line options listed below and/or the 'mv-rm-files' setting.
+**
+** The 'rename' command never renames or moves files on disk, even when the
+** command line options and/or the 'mv-rm-files' setting would otherwise
+** require it to do so.
+**
+** WARNING: If the "--hard" option is specified -OR- the "mv-rm-files"
+**          setting is non-zero, files WILL BE renamed or moved on disk
+**          as well.  This does NOT apply to the 'rename' command.
 **
 ** Options:
+**   --soft                  Skip moving files within the checkout.
+**                           This supersedes the --hard option.
+**   --hard                  Move files within the checkout.
 **   --case-sensitive <BOOL> Override the case-sensitive setting.
+**   -n|--dry-run            If given, display instead of run actions.
 **
 ** See also: changes, status
 */
 void mv_cmd(void){
   int i;
   int vid;
+  int moveFiles;
+  int dryRunFlag;
+  int softFlag;
+  int hardFlag;
   char *zDest;
   Blob dest;
   Stmt q;
 
   db_must_be_within_tree();
+  dryRunFlag = find_option("dry-run","n",0)!=0;
+  softFlag = find_option("soft",0,0)!=0;
+  hardFlag = find_option("hard",0,0)!=0;
 
   /* We should be done with options.. */
   verify_all_options();
@@ -653,6 +843,19 @@ void mv_cmd(void){
   }
   zDest = g.argv[g.argc-1];
   db_begin_transaction();
+  if( g.argv[1][0]=='r' ){ /* i.e. "rename" */
+    moveFiles = 0;
+  }else if( softFlag ){
+    moveFiles = 0;
+  }else if( hardFlag ){
+    moveFiles = 1;
+  }else{
+#if FOSSIL_ENABLE_LEGACY_MV_RM
+    moveFiles = db_get_boolean("mv-rm-files",0);
+#else
+    moveFiles = FOSSIL_MV_RM_FILE;
+#endif
+  }
   file_tree_name(zDest, &dest, 1);
   db_multi_exec(
     "UPDATE vfile SET origname=pathname WHERE origname IS NULL;"
@@ -711,10 +914,12 @@ void mv_cmd(void){
   while( db_step(&q)==SQLITE_ROW ){
     const char *zFrom = db_column_text(&q, 0);
     const char *zTo = db_column_text(&q, 1);
-    mv_one_file(vid, zFrom, zTo);
+    mv_one_file(vid, zFrom, zTo, dryRunFlag);
+    if( moveFiles ) add_file_to_move(zFrom, zTo);
   }
   db_finalize(&q);
   db_end_transaction(0);
+  if( moveFiles ) process_files_to_move(dryRunFlag);
 }
 
 /*
